requirements.txt
numpy
opencv-python
pillow
youtube-dl
# youtube-upload
git+https://github.com/tokland/youtube-upload.git